requirements.txt
openai
faiss-cpu
streamlit
nltk